Reorganized code, added InitializingAssignmentsEvent
